rename: 끓인 팟 form 컴포넌트 이름 변경
